backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
httpx>=0.25.0
google-cloud-bigquery>=3.0.0
python-dotenv>=1.0.0